test:test: consolidate variable-sort helpers into actions.ts
