+++ tests/utils/page/actions.ts
@@ -846,6 +846,31 @@ const addRowToActiveTab = async (
   });
 };
 
+// Clicks the sort toggle once: default -> asc -> desc -> default.
+const cycleVariableSort = async (page: Page) => {
+  await test.step('Cycle variable sort mode', async () => {
+    await buildCommonLocators(page).environment.sortToggle().click();
+  });
+};
+
+// Displayed variable names, top to bottom, excluding the trailing empty row.
+const getVisibleVariableNames = async (page: Page): Promise<string[]> => {
+  const values = await buildCommonLocators(page)
+    .environment.visibleNameInputs()
+    .evaluateAll((inputs) => inputs.map((el) => (el as HTMLInputElement).value));
+  return values.filter((name) => name !== '');
+};
+
+// Drags row `fromName` onto row `toName` (Manual sort mode only).
+const dragVariableRow = async (page: Page, fromName: string, toName: string) => {
+  await test.step(`Drag variable row "${fromName}" onto "${toName}"`, async () => {
+    const common = buildCommonLocators(page);
+    await common.environment.varRow(fromName).dragTo(common.environment.varRow(toName), {
+      targetPosition: { x: 5, y: 5 }
+    });
+  });
+};
+
 /**
  * Delete every global environment in the workspace. Global environments persist at
  * the workspace level (closeAllCollections does not remove them), so call this to keep
@@ -2531,6 +2556,9 @@ export {
   addEnvironmentVariable,
   addEnvironmentVariables,
   addRowToActiveTab,
+  cycleVariableSort,
+  getVisibleVariableNames,
+  dragVariableRow,
   deleteAllGlobalEnvironments,
   saveEnvironment,
   closeEnvironmentPanel,
